+++ doc/changelog.rst
@@ -8,6 +8,9 @@ Version 1.1
 * ``Runner.bench_command()`` now measures also the maximum RSS memory if
   available.
 * Fix Windows 32bit issue on Python 2.7, fix by yattom.
+* ``Runner.bench_func()`` now uses ``functools.partial()`` if the function
+  has argument. Calling ``partial()`` is 1.07x faster (-6%) than calling
+  ``func(*args)``.
 
 Version 1.0 (2017-03-17)
 ------------------------